Air resistance approximation
Added a simulation for air resistance. This effect is generally quite small except on short, light, or very fast moving trains. For example, a 10 car fully loaded frieght train running at 95 km/hr had a wind resistance about 40% of its rolling resistance.

This simulation makes a lot of assumptions due to lack of information. In particular, it assumes all cars of the train have the same cross-sectional area in the same shape. In a train with varying cars, the difference in cross section from one car to the next would increase air resistance. In trains with smaller cars, air resistance would be proportionally smaller. Overall I think the approximation is within a factor of 2 of the real value in most cases.