Inheritance Problem
I'd like this to make a new multiblock that inherits all the functions of the RailRollerMultiblock, but with a new set of blocks and offsets. What this actually does is create a new multiblock that does not tick, and would crash if it did, because its offsets are not working across the inheritance barrier.